ACP Proyecciones/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sse-starlette>=1.8.2
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
pyodbc
python-dotenv
scalar-fastapi
python-multipart
pandas
numpy
openpyxl
rapidfuzz
python-jose>=3.3.0
cryptography>=40.0.0
passlib>=1.7.4
bcrypt>=4.0.0
pytest>=8.0.0
streamlit>=1.30.0
requests>=2.31.0
